[ctrl_flow_struct] Rename tests to test
